pyproject.toml
[project]
name = "md2office"
version = "0.1.0"
description = "Convert Markdown to professional DOCX documents with corporate template support"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.12"
authors = [
    {name = "md2office contributors"}
]
keywords = ["markdown", "docx", "converter", "word", "document"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Framework :: Litestar",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Documentation",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Typing :: Typed",
]
dependencies = [
    "litestar[standard]>=2.0",
    "python-docx>=1.0",
    "mistune>=3.0",
    "msgspec>=0.18",
    "jinja2>=3.1",
    "cyclopts>=3.0",
    "pyyaml>=6.0",
    "aiofiles>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "httpx>=0.28",
    "ruff>=0.8",
]

[project.scripts]
md2office = "md2office.cli:main"

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/md2office"
Documentation = "https://github.com/YOUR_USERNAME/md2office#readme"
Repository = "https://github.com/YOUR_USERNAME/md2office.git"
Issues = "https://github.com/YOUR_USERNAME/md2office/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/md2office"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict
]

[tool.ruff.lint.isort]
known-first-party = ["md2office"]

[tool.coverage.run]
source = ["src/md2office"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
